requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9

# Cloud Services
boto3==1.29.7
replicate==0.22.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Image Processing (for future face swap preprocessing)
Pillow==10.1.0

# Training & Model Hub
huggingface-hub==0.20.0
